+++ dropwizard-config-updater/src/test/groovy/ru/vyarus/yaml/updater/dropwizard/ErrorsTest.groovy
@@ -9,9 +9,9 @@ class ErrorsTest extends AbstractTest {
     def "Check invalid file url"() {
 
         expect: "invalid file url"
-        def res = runWithError("config.yml", "http://localhost/file.yml")
+        def res = runWithError("config.yml", "http://localhost:2222/file.yml")
         // different on lin and win
-        res.contains("Update file not found: http://localhost/file.yml") || res.contains("Failed to load file from url: http://localhost/file.yml")
+        res.contains("Update file not found: http://localhost:2222/file.yml") || res.contains("Failed to load file from url: http://localhost:2222/file.yml")
 
     }