+++ yaml-config-updater/src/test/groovy/ru/vyarus/yaml/updater/parse/struct/TreePathsTest.groovy
@@ -63,18 +63,20 @@ class TreePathsTest extends Specification {
         toPaths(tree) == """   1| simple
    4| quoted
    7| quoted2
-  10| include_newlines
-  16| middle_newlines
-  24| fold_newlines
-  30| ignore_ind
-  35| append_ind
-  40| custom_indent
-  47| custom_indent2
-  53| object/sub
-  58| list[0]
-  60| list[1]/obj
-  62| list[2]/ob2
-  66| flow"""
+  10| multiline
+  14| multiline2
+  19| include_newlines
+  25| middle_newlines
+  33| fold_newlines
+  39| ignore_ind
+  44| append_ind
+  49| custom_indent
+  56| custom_indent2
+  62| object/sub
+  67| list[0]
+  69| list[1]/obj
+  71| list[2]/ob2
+  75| flow"""
     }
 
     private String toPaths(StructTree tree) {